moving half baked code to new computer
